delete user by id implemented. merged test and testrequest controllers
