+++ src/main/java/com/intern_project/test_management_service/controllers/TestController.java
@@ -1,6 +1,8 @@
 package com.intern_project.test_management_service.controllers;
 
 import com.intern_project.test_management_service.models.Test;
+import com.intern_project.test_management_service.models.TestRequest;
+import com.intern_project.test_management_service.services.TestRequestService;
 import com.intern_project.test_management_service.services.TestService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -14,6 +16,8 @@ public class TestController {
 
     private final JdbcTemplate jdbcTemplate;
     private final TestService testService;
+    private final TestRequestService testRequestService;
+
 
     @GetMapping("/getTests")
     public List<Test> getTests() {
@@ -27,10 +31,21 @@ public class TestController {
         return testService.addTest(test);
     }
 
-    @DeleteMapping("/deleteTestsAndResetSequence")
-    public void deleteAllTests() {
+    @GetMapping("/getTestRequests")
+    public List<TestRequest> getTestRequests() {
 
-        String sqlTruncate = "TRUNCATE TABLE tests RESTART IDENTITY";
-        jdbcTemplate.update(sqlTruncate);
+        return testRequestService.getTestRequests();
+    }
+
+    @GetMapping("/getTestsRequestsId")
+    public List<TestRequest> getTestsId(@RequestBody Long userId) {
+
+        return testRequestService.getTests(userId);
+    }
+
+    @PostMapping("/createTestRequest")
+    public TestRequest createTestRequest(@RequestBody TestRequest testRequest) {
+
+        return testRequestService.addTestRequest(testRequest);
     }
 }
