+++ src/main/java/com/intern_project/test_management_service/controllers/UserController.java
@@ -1,11 +1,8 @@
 package com.intern_project.test_management_service.controllers;
 
-import com.intern_project.test_management_service.models.Test;
-import com.intern_project.test_management_service.models.TestType;
 import com.intern_project.test_management_service.models.User;
 import com.intern_project.test_management_service.services.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
 
@@ -36,4 +33,17 @@ public class UserController {
         String sqlTruncate = "TRUNCATE TABLE users RESTART IDENTITY CASCADE";
         jdbcTemplate.update(sqlTruncate);
     }
+
+    @DeleteMapping("/deleteUser/{id}")
+    public void deleteUserById(@PathVariable Long id) {
+        // Delete user roles associated with the user
+        String deleteUserRolesSql = "DELETE FROM user_roles WHERE user_id = ?";
+        jdbcTemplate.update(deleteUserRolesSql, id);
+
+        // Delete the user
+        String deleteUserSql = "DELETE FROM users WHERE id = ?";
+        jdbcTemplate.update(deleteUserSql, id);
+    }
+
+
 }
